scraper/requirements.txt
selenium==4.28.1
webdriver_manager==4.0.2
pillow==11.2.1
pytest==8.3.4
requests==2.32.3